update usb interface wand add watcher thread to get device status

+++ jkinterface/api_libcups.h
@@ -41,7 +41,7 @@ typedef struct Job_struct{
     int pages_completed;
     int copies;
     int state;
-    long timet;
+    long long timet;
 
 }
     Job_struct;
